+++ tests/killNode.R
@@ -11,6 +11,9 @@ options(parallelly.debug = FALSE)
 
 message("*** killNode() and isNodeAlive() ...")
 
+isNodeAliveSupported <- isTRUE(parallelly:::pid_exists(Sys.getpid()))
+message("isNodeAlive() works: ", isNodeAliveSupported)
+
 cl <- makeClusterPSOCK(2L, autoStop = FALSE)
 names(cl) <- sprintf("Node %d", seq_along(cl))
 print(cl)
@@ -37,11 +40,15 @@ alive <- isNodeAlive(cl)
 print(alive)
 stopifnot(
   length(alive) == length(cl),
-  is.logical(alive),
-  !anyNA(alive),
-  isTRUE(alive[[1]]), isTRUE(alive[[2]]),
-  all(alive)
+  is.logical(alive)
 )
+if (isNodeAliveSupported) {
+  stopifnot(
+    !anyNA(alive),
+    isTRUE(alive[[1]]), isTRUE(alive[[2]]),
+    all(alive)
+  )
+}
 
 message("- Terminate cluster nodes")
 signaled <- killNode(cl)
@@ -77,6 +84,7 @@ repeat {
 
 ## Remove any stray Rscript<hexcode> files
 if (.Platform$OS.type == "windows") {
+  if (!isNodeAliveSupported) Sys.sleep(5.0)
   tmpfiles <- tmpfiles[utils::file_test("-f", tmpfiles)]
   if (length(tmpfiles) > 0L) {
     warning(sprintf("Cleaning up temporary left-over files: [n=%d] %s",