+++ NEWS.md
@@ -31,7 +31,7 @@
    limit the total CPU load, is with Linux containers, e.g. `docker
    run --cpus=3.5 ...`.
 
- * Add support for `availableCores(method = "connections")`, which
+ * Add support for `availableCores(methods = "connections")`, which
    returns the current number of free R connections per
    `freeConnections()`.  This is the maximum number of PSOCK, SOCK,
    and MPI **parallel** cluster nodes we can open without running out