+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version (development version)
+# Version 1.33.0 [2022-12-13]
 
 ## New Features